pyproject.toml
[project]
name = "mvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate==0.33.0",
    "faster-whisper==1.0.1",
    "ffmpeg-python==0.2.0",
    "flask>=3.1.1",
    "fpdf==1.7.2",
    "gtts>=2.5.4",
    "langchain-google-genai>=2.1.9",
    "openai>=1.99.9",
    "opencv-python>=4.11.0.86",
    "pyaudio>=0.2.14",
    "pydub==0.25.1",
    "python-dotenv>=1.1.1",
    "speechrecognition>=3.14.3",
    "streamlit==1.38.0",
    "streamlit-javascript>=0.1.5",
    "streamlit-webrtc==0.47.1",
    "torch>=2.2.0",
    "transformers==4.36.2",
    "werkzeug>=3.1.3",
    "whisper>=1.1.10",
]
